+++ CHANGELOG.md
@@ -1,3 +1,5 @@
-## 1.0.1
+## 1.1.0
 - Animation is now way smoother
 - Reduced amount of packets sent by 50%
+- Added UpdateChecker (can be configured in update-checker.yml)
+- Added Discord verification (only needed if you need support on Discord)